Feat/add ci (#19)
* chore: Update gitleaks to version v8.18.4

* feat: Update AI Studio compute instance module with assigned user and tenant

This commit updates the AI Studio compute instance module to include the `assignedUserId` and `assignedUserTenant` parameters. These parameters specify who the compute instance is assigned to, allowing only them to access it. This enhancement improves access control and security for AI Studio resources.

* feat: Update AI Studio compute instance module with assigned user and tenant

* feat: Update AI Studio compute instance module with VPN support

This commit adds support for VPN configuration in the AI Studio compute instance module. It introduces new parameters `enablevpn`, `vpnVnetName`, and `vpnVnetResourceGroupName` to enable and configure VPN connectivity for the compute instance. This enhancement enhances the network security and flexibility of AI Studio resources.